# constants: add RTSP/RAOP methods