Supersedes: #66 

cc @nodejs/http