@indutny 2020-10-16
Benchmarks:

**with** change:
```
http loose: "seanmonstar/httparse" (C)
8192.00 mb | 1529.18 mb/s | 2280879.50 ops/sec | 5.36 s
http strict: "seanmonstar/httparse" (C)
8192.00 mb | 1414.58 mb/s | 2109943.20 ops/sec | 5.79 s
http loose: "nodejs/http-parser" (C)
8192.00 mb | 1044.18 mb/s | 2105571.59 ops/sec | 7.85 s
http strict: "nodejs/http-parser" (C)
8192.00 mb | 1004.79 mb/s | 2026150.04 ops/sec | 8.15 s
```

**without** change:
```
http loose: "seanmonstar/httparse" (C)
8192.00 mb | 1559.15 mb/s | 2325589.03 ops/sec | 5.25 s
http strict: "seanmonstar/httparse" (C)
8192.00 mb | 1402.73 mb/s | 2092269.69 ops/sec | 5.84 s
http loose: "nodejs/http-parser" (C)
8192.00 mb | 1081.31 mb/s | 2180449.94 ops/sec | 7.58 s
http strict: "nodejs/http-parser" (C)
8192.00 mb | 1027.44 mb/s | 2071824.99 ops/sec | 7.97 s
```