@pallas 2020-10-16
After this change, I'm getting HPE_INVALID_CONSTANT again.. going to investigate why.