@pallas 2020-10-16
Ah, I see my current problem. It looks like Apple sends GETs in addition to the listed verbs.