@pallas 2020-10-16
Ok, can you please add GET and POST to the RTSP verbs? After doing so, this works great with AirPlay.