@pallas 2020-10-18
Verified that this works perfectly for AirPlay.