@indutny 2020-10-18
...and released as [2.2.0](https://github.com/nodejs/llhttp/releases/tag/v2.2.0)